Keep disk space from blowing up
Do the deletions inside the loop instead of after it. The project goal isn't speed; it's space management. If we wait until the end, we risk running out of local storage before we're done. If we delete as we go, we get the same net effect without needing space for everything.

+++ rainbow_compression.ps1
@@ -20,12 +20,14 @@ ForEach ($compression_algorithm In $compression_algorithms) {
         $output_file="$($input_file)_$($i).7z"
         7z a "-mx=$($compression_level)" "-mf=$($executable_compression)" "-mhc=$($header_compression)" "-m0=$($compression_algorithm)" "$($output_file)" "$($input_file)"
         $i=$i+1
+        If ($i -Ge 2) {
+          # Doing the deletion inside the loop slows it down, but it keeps the disk space in check
+          $output_files = Get-ChildItem "$($input_file)_*.7z" | Sort-Object Length
+          $file_to_keep,$files_to_delete=$output_files
+          $file_to_keep | Rename-Item -NewName "$($input_file).7z"
+          $files_to_delete | Remove-Item
+        }
       }
     }
   }
 }
-
-$output_files = Get-ChildItem "$($input_file)_*.7z" | Sort-Object Length
-$file_to_keep,$files_to_delete=$output_files
-$file_to_keep | Rename-Item -NewName "$($input_file).7z"
-$files_to_delete | Remove-Item
